rerange files and function
